api/requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
pydantic>=2.0
pydantic-settings>=2.0

# HTTP client
httpx>=0.27.0

# Embeddings
sentence-transformers>=3.0.0

# Qdrant
qdrant-client>=1.7.0

# Document processing
docling>=2.0.0
python-docx>=1.0.0
python-pptx>=3.0.0
pytesseract>=0.3.10
Pillow>=10.0.0

# Web scraping
beautifulsoup4>=4.12.0
requests>=2.32.0

# Utilities
python-dotenv>=1.0.0

# Evaluation (optional - install separately if needed)
# ragas>=0.1.0
# datasets>=2.14.0